Изменение поля double на decimal в WPF

+++ DiscRental73TestWpf/Infrastructure/DialogWindowServices/Strategies/ShowProductCostStrategy.cs
@@ -12,8 +12,8 @@ namespace DiscRental73TestWpf.Infrastructure.DialogWindowServices.Strategies
     {
         #region Ограничения на ввод данных 
 
-        public double CostMaxValue { get; set; }
-        public double CostMinValue { get; set; }
+        public decimal CostMaxValue { get; set; }
+        public decimal CostMinValue { get; set; }
 
         #endregion
