+++ DiscRental73TestWpf/Infrastructure/HelperModels/IssueRentalBindingModel.cs
@@ -8,6 +8,6 @@ namespace DiscRental73TestWpf.Infrastructure.HelperModels
         public int ClientId { get; set; }
         public DateTime DateOfIssue { get; set; }
         public DateTime DateOfReturn { get; set; }
-        public double PledgeSum { get; set; }
+        public decimal PledgeSum { get; set; }
     }
 }
