+++ DiscRental73TestWpf/Infrastructure/HelperModels/IssueReturnBindingModel.cs
@@ -3,6 +3,6 @@
     public class IssueReturnBindingModel
     {
         public int RentalId { get; set; }
-        public double ReturnSum { get; set; }
+        public decimal ReturnSum { get; set; }
     }
 }
